collector/requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==24.2.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
cattrs==24.1.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
docstring-to-markdown==0.15
exceptiongroup==1.2.2
fastapi==0.115.11
fonttools==4.56.0
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
idna==3.10
importlib-metadata==6.7.0
jedi==0.19.2
jedi-language-server==0.42.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lsprotocol==2023.0.1
MarkupSafe==3.0.2
matplotlib==3.10.1
microvenv==2023.2.0
mplfinance==0.12.10b0
multitasking==0.0.11
narwhals==1.31.0
numpy==2.2.4
packaging==23.1
pandas==2.2.3
parso==0.8.4
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.7
protobuf==5.29.4
psycopg2-binary==2.9.10
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pygls==1.3.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
streamlit==1.43.2
tenacity==9.0.0
toml==0.10.2
tomli==2.0.1
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
yfinance==0.2.55
zipp==3.15.0
